requirements.txt
dvc
numpy
pandas
scikit-learn
kaggle
feature_engine
matplotlib
seaborn
python-dotenv
streamlit